requirements.txt
configparser==6.0.0
paho-mqtt==1.6.1
